DDSim: move UI commands to separate helper to give better documentation for it
